+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## **1.1.0**
+
+---
+
+* Added `ComputeLua.CreateThead(actor: Actor, threadName: name, callback: (number, VariableBufferDataType) -> ())`
+* Changed `Dispatcher:Dispatch()` to return a `Promise.defer()`
+* Variable Buffer is now a SharedTable
+
 ## **1.0.0**
 
 ---